fix(SPE-31a): review harden opportunity card and tests
- Factor shared courier opportunity fields in buildCourierNetworkCapacityOpportunityCard
- Use stable React keys for mitigation label chips (label-index)
- Centralize withPaidCourierAndFunding in src/test/fixtures for hub and gap tests
